+++ CHANGELOG.md
@@ -56,6 +56,8 @@ Bug fixes:
 * Made `rb_respond_to` work with primitives (#1869, @chrisseaton).
 * Fixed issue with missing backtrace for `rescue $ERROR_INFO` (#1660).
 * Fixed `Struct#hash` for `keyword_init: true` `Struct`.
+* Fixed `String#{upcase!,downcase!,swapcase!}(:ascii)` for non-ASCII-compatible encodings like UTF-16.
+* Fixed `String#capitalize!` for strings that weren't full ASCII.
 
 Compatibility: